Move the pixel traits to a stand alone crate v_frame
