+++ Cargo.toml
@@ -68,6 +68,7 @@ dav1d-sys = { version = "0.3", optional = true }
 aom-sys = { version = "0.1.3", optional = true }
 scan_fmt = { version = "0.2.3", optional = true, default-features = false }
 ivf = { version = "0.1", path = "ivf/", optional = true }
+v_frame = { version = "0.1", path = "v_frame/" }
 rayon = "1.0"
 toml = { version = "0.5", optional = true }
 arrayvec = "0.5"
@@ -149,7 +150,7 @@ codegen-units = 1
 lto = true
 
 [workspace]
-members = [".", "ivf"]
+members = [".", "ivf", "v_frame"]
 
 [package.metadata.docs.rs]
 no-default-features = true
